Moved generateResourceMessage to a genericAttachmentsTemplate under /data
